Work-in-progress unit tests
